+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - Issue 203 - Construct CNM to trigger load data operations and ingest granule
     - Issue 236 - Allow UAT query of CMR to support querying in different venues
 ### Changed
+    - Issue 251 - Add note to readme to point to documentation
 ### Deprecated
 ### Removed
 ### Fixed