+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated
 ### Removed
 ### Fixed
+    - Issue 258 - Granules with very large feature counts cannot be added to hydrocron
     - Issue 235 - Track ingest table can be populated with granules that aren't loaded into Hydrocron
     - Issue 248 - Track ingest operations need to query UAT for granule files if track ingest is running in SIT or UAT
 ### Security